requirements.txt
# Core dependencies
sqlite3  # Built-in Python module for database
dataclasses  # Built-in Python module for data models

# Vector search and embeddings
faiss-cpu>=1.7.0
numpy>=1.21.0

# LLM API clients
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# HTTP requests
requests>=2.28.0

# CLI interface
click>=8.0.0

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.991